添加了form相关功能

+++ langmovie/src/main/java/cn/langlang/langmovie/dao/UserDao.java
@@ -17,7 +17,7 @@ public interface UserDao {
      * @param user_id
      * @return TestUser
      */
-    User getUserById(Integer user_id);
+    User getUserById(Long user_id);
 
     /**
      *
